pyproject.toml
[project]
name = "nl2ml-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "adjusttext>=1.3.0",
    "beautifulsoup4>=4.13.5",
    "brokenaxes>=0.6.2",
    "google-cloud-bigquery>=3.25.0",
    "hf-xet>=1.1.8",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "langchain>=0.3.27",
    "langchain-community>=0.3.21",
    "langchain-google-genai>=2.1.10",
    "langchain-openai>=0.3.32",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "mlflow>=2.22.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "seaborn>=0.13.2",
    "setuptools>=80.7.1",
    "sqlglot>=27.10.0",
    "torch==2.5.1",
    "unsloth==2025.3.14",
    "vllm==0.7.3",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
